pyproject.toml
[tool.poetry]
name = "sentry"
version = "0.1.0"
description = "Send configuration to device via serial."
authors = ["ikievite <ikievite@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
mypy = "^0.971"
wemake-python-styleguide = "^0.16.1"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sentry = "sentry.scripts.sentry:main"
